refactor(interpreter): split declarations destructuring and integrity
